refactor:#99 getUserProfile 함수 일반 서버 함수로 이동 - 서버 액션 파일에서 해당 함수 삭제 - 바뀐 함수 적용 - 클라이언트에서 사용하기 위한 라우트 핸들러 생성
